+++ apps/studio/components/interfaces/Settings/API/DataApiProjectUrlCard.tsx
@@ -17,9 +17,9 @@ import { ShimmeringLoader } from 'ui-patterns/ShimmeringLoader'
 
 import { getApiEndpoint } from '@/components/interfaces/Integrations/DataApi/DataApi.utils'
 import { DatabaseSelector } from '@/components/ui/DatabaseSelector'
+import { useProjectApiUrl } from '@/data/config/project-endpoint-query'
 import { useLoadBalancersQuery } from '@/data/read-replicas/load-balancers-query'
 import { useReadReplicasQuery } from '@/data/read-replicas/replicas-query'
-import { useProjectApiUrl } from '@/hooks/misc/useProjectApiUrl'
 import { useSelectedProjectQuery } from '@/hooks/misc/useSelectedProject'
 import { useStaticEffectEvent } from '@/hooks/useStaticEffectEvent'
 import { useDatabaseSelectorStateSnapshot } from '@/state/database-selector'
